+++ loader/app/index.ts
@@ -1,8 +1,15 @@
-import * as loader from "./targets/app";
-import * as loader_base from "./loader/loader";
-window["loader"] = loader_base;
-    /* let the loader register himself at the window first */
-setTimeout(loader.run, 0);
+import "core-js/stable";
+import "./polifill";
+
+import * as loader from "./loader/loader";
+window["loader"] = loader;
+/* let the loader register himself at the window first */
+
+import * as AppLoader from "./targets/app";
+setTimeout(AppLoader.run, 0);
+
+import * as EmptyLoader from "./targets/empty";
+//setTimeout(EmptyLoader.run, 0);
 
 export {};