+++ loader/app/loader/utils.ts
@@ -10,11 +10,7 @@ export class LoadSyntaxError {
 
 export function script_name(path: SourcePath, html: boolean) {
     if(Array.isArray(path)) {
-        let buffer = "";
-        let _or = " or ";
-        for(let entry of path)
-            buffer += _or + script_name(entry, html);
-        return buffer.slice(_or.length);
+        return path.filter(e => !!e).map(e => script_name(e, html)).join(" or ");
     } else if(typeof(path) === "string")
         return html ? "<code>" + path + "</code>" : path;
     else
@@ -35,31 +31,40 @@ export interface ParallelResult<T> {
     skipped: T[];
 }
 
-export async function load_parallel<T>(requests: T[], executor: (_: T) => Promise<void>, stringify: (_: T) => string, options: ParallelOptions) : Promise<ParallelResult<T>> {
+export type LoadCallback<T> = (entry: T, state: "loading" | "loaded") => void;
+
+export async function load_parallel<T>(requests: T[], executor: (_: T) => Promise<void>, stringify: (_: T) => string, options: ParallelOptions, callback?: LoadCallback<T>) : Promise<ParallelResult<T>> {
     const result: ParallelResult<T> = { failed: [], succeeded: [], skipped: [] };
-    const pending_requests = requests.slice(0).reverse(); /* we're only able to pop from the back */
-    const current_requests = {};
+    const pendingRequests = requests.slice(0).reverse(); /* we're only able to pop from the back */
+    const currentRequests = {};
 
-    while (pending_requests.length > 0) {
-        while(typeof options.max_parallel_requests !== "number" || options.max_parallel_requests <= 0 || Object.keys(current_requests).length < options.max_parallel_requests) {
-            const script = pending_requests.pop();
-            const name = stringify(script);
+    if(typeof callback === "undefined")
+        callback = () => {};
 
-            current_requests[name] = executor(script).catch(e => result.failed.push({ request: script, error: e })).then(() => {
-                delete current_requests[name];
+    options.max_parallel_requests = 1;
+    while (pendingRequests.length > 0) {
+        while(typeof options.max_parallel_requests !== "number" || options.max_parallel_requests <= 0 || Object.keys(currentRequests).length < options.max_parallel_requests) {
+            const element = pendingRequests.pop();
+            const name = stringify(element);
+
+            callback(element, "loading");
+            currentRequests[name] = executor(element).catch(e => result.failed.push({ request: element, error: e })).then(() => {
+                delete currentRequests[name];
+                callback(element, "loaded");
             });
-            if(pending_requests.length == 0) break;
+            if(pendingRequests.length == 0)
+                break;
         }
 
         /*
          * Wait 'till a new "slot" for downloading is free.
          * This should also not throw because any errors will be caught before.
          */
-        await Promise.race(Object.keys(current_requests).map(e => current_requests[e]));
+        await Promise.race(Object.keys(currentRequests).map(e => currentRequests[e]));
         if(result.failed.length > 0)
             break; /* finish loading the other requests and than show the error */
     }
-    await Promise.all(Object.keys(current_requests).map(e => current_requests[e]));
-    result.skipped.push(...pending_requests);
+    await Promise.all(Object.keys(currentRequests).map(e => currentRequests[e]));
+    result.skipped.push(...pendingRequests);
     return result;
 }